+++ src/devteam/tools/__init__.py
@@ -1,5 +1,7 @@
 from .sandbox import DockerSandbox
+from .registry import tool_registry
 
 __all__ = [
-    'DockerSandbox'
+    'DockerSandbox',
+    'tool_registry'
 ]